+++ store/src/main/java/com/mallfoundry/store/product/repository/jpa/JpaProductRepository.java
@@ -23,5 +23,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface JpaProductRepository
-        extends ProductRepository, JpaRepository<Product, Integer> {
+        extends ProductRepository, JpaRepository<Product, Long> {
 }
