+++ store/src/main/java/com/mallfoundry/store/rest/ProductResourceV1.java
@@ -16,8 +16,8 @@
 
 package com.mallfoundry.store.rest;
 
-import com.mallfoundry.store.product.ProductService;
 import com.mallfoundry.store.product.Product;
+import com.mallfoundry.store.product.ProductService;
 import com.mallfoundry.store.product.search.ProductQuery;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -43,7 +43,7 @@ public class ProductResourceV1 {
     }
 
     @GetMapping("/products/{id}")
-    public Optional<Product> getProduct(@PathVariable("id") Integer id) {
+    public Optional<Product> getProduct(@PathVariable("id") Long id) {
         return this.productService.getProduct(id);
     }
 }
